testing/web-platform/tests/workers/WorkerLocation_hash_encoding.htm

<!DOCTYPE html>
<html>
<head>
<title> WorkerLocation.hash with url encoding string </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    
    var FileName = "./support/WorkerLocation.js";
    var HashString = "#question%3f";
    var ExpectedResult = HashString;
    
    var t = async_test("Test Description: WorkerLocation.hash returns '#question%3f' from input 'http://example.com/carrot#question%3f'");
                  
    var worker = new Worker(FileName + HashString);

    worker.onmessage = t.step_func(function(evt)
    {
        assert_equals(evt.data.hash, ExpectedResult);
        t.done();
    });
    
</script>
</body>
</html>
